chore: moves utilities out of components folder
