+++ packages/appChrome/components/AppChrome.tsx
@@ -8,7 +8,7 @@ import {
   flush
 } from "../../shared/styles/styleUtils";
 import { ThemeProvider } from "emotion-theming";
-import getCSSVarValue from "../../utilities/components/getCSSVarValue";
+import { getCSSVarValue } from "../../utilities";
 import { themeBgPrimaryInverted } from "../../design-tokens/build/js/designTokens";
 import {
   defaultSidebarItemHorizPaddingSize,
