+++ packages/appChrome/components/HeaderBar.tsx
@@ -9,7 +9,7 @@ import {
   themeTextColorPrimaryInverted
 } from "../../design-tokens/build/js/designTokens";
 import { pickReadableTextColor } from "../../shared/styles/color";
-import getCSSVarValue from "../../utilities/components/getCSSVarValue";
+import { getCSSVarValue } from "../../utilities";
 
 export interface HeaderProps {
   children: React.ReactElement<HTMLElement> | string;
