+++ packages/utilities/index.ts
@@ -1,7 +1,5 @@
-export { default as min } from "./components/min";
-export { default as max } from "./components/max";
-export { default as percentage } from "./components/percentage";
-export { default as compose } from "./components/compose";
 export {
   default as findNestedPropertyInObject
 } from "./findNestedPropertyInObject";
+export { default as getCSSVarValue } from "./getCSSVarValue";
+export { default as resizeEventManager } from "./resizeEventManager";
