+++ packages/utilities/interactionMode.ts
