+++ packages/utilities/max.ts
